🚸 Reverse event order in GetLastEvents query to show the most recent events first
Modified the SQL ORDER BY clause in the GetLastEvents function of the EventRepository to order events by started_at in descending order instead of ascending order. This change will display the most recently started events first, better aligning with the common expectations of 'last events' functionality. This adjustment enhances the usability and relevance of query results when retrieving a limited set of recent events.